feat:NRC20 Contract Inner Transfer Info Rollback

+++ src/main/java/io/nuls/api/entity/ContractTokenTransferInfo.java
@@ -2,6 +2,8 @@ package io.nuls.api.entity;
 
 import io.nuls.api.model.ContractTokenTransferDto;
 
+import java.math.BigInteger;
+
 public class ContractTokenTransferInfo {
     private String txHash;
 
@@ -9,7 +11,7 @@ public class ContractTokenTransferInfo {
 
     private String toAddress;
 
-    private Long txValue;
+    private BigInteger txValue;
 
     private Long createTime;
 
@@ -20,7 +22,7 @@ public class ContractTokenTransferInfo {
     public ContractTokenTransferInfo(ContractTokenTransferDto dto) {
         this.fromAddress = dto.getFrom();
         this.toAddress = dto.getTo();
-        this.txValue = dto.getValue().longValue();
+        this.txValue = dto.getValue();
     }
 
     public String getContractAddress() {
@@ -55,11 +57,11 @@ public class ContractTokenTransferInfo {
         this.toAddress = toAddress == null ? null : toAddress.trim();
     }
 
-    public Long getTxValue() {
+    public BigInteger getTxValue() {
         return txValue;
     }
 
-    public void setTxValue(Long txValue) {
+    public void setTxValue(BigInteger txValue) {
         this.txValue = txValue;
     }
